+++ CHANGELOG.md
@@ -25,6 +25,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed misleading error message when digest header is not signed.
 - Fixed double-decoding of media URL in media proxy view.
 
+### Security
+
+- Verify permissions on embedded objects when processing activities sent to FEP-ae97 outbox.
+
 ## [5.3.0] - 2026-05-13
 
 ### Added